+++ docs/changes.rst
@@ -8,6 +8,7 @@ Released on ? 2015.
 
 * Dropped support for Django 1.6 and South migrations.
 * Support for adding new translations when using Java Property files
+* Allow to accept suggestion without editing.
 
 weblate 2.2
 -----------